# Perf: Optimize BlockFinder by reducing RPC calls