## Summary
- Identified and optimized the slowest test without using mocks
- Reduced RPC calls by implementing smarter search algorithms
- Achieved 30% performance improvement on the slowest test

## Test Plan
- [x] Run all existing tests - all pass
- [x] Verify performance improvements - 30% faster on slowest test
- [x] Ensure no mocks were introduced - only real RPC calls used
- [x] Run lint and typecheck - all pass

## Results
- Individual test: 11.8s → 8.3s (30% improvement)
- Full test suite: 38.1s → 30.6s (20% improvement)

Resolves #120